TIFFFormat: change writing of multi-channel images
Changes the way multi-channel TIFFs are written. According to the TIFF6
specification, RGB images can be saved with multiple samples per pixel.
That is, an image with 3 channels is written into one TIFF plane with
the values of the 3 channels in one sample.

In contrast to that, ImageJ1 writes each channel into an individual
plane with only one sample per pixel, reconstructing the hyperstack from
the metadata in the header. This commit changes the saving scheme to
conform with ImageJ1.